Sub levels work. Move around some parsing stuff.
