+++ src/IKVM.Tools.Importer/CompilerClassLoader.cs
@@ -3261,317 +3261,4 @@ namespace IKVM.Tools.Importer
         }
     }
 
-    sealed class Jar
-    {
-        internal readonly string Name;
-        private readonly List<JarItem> Items = new List<JarItem>();
-
-        internal Jar(string name)
-        {
-            this.Name = name;
-        }
-
-        internal Jar Copy()
-        {
-            Jar newJar = new Jar(Name);
-            newJar.Items.AddRange(Items);
-            return newJar;
-        }
-
-        internal void Add(string name, byte[] data, FileInfo fileInfo = null)
-        {
-            Items.Add(new JarItem(name, data, fileInfo));
-        }
-
-        private readonly struct JarItem
-        {
-            internal readonly string name;
-            internal readonly byte[] data;
-            internal readonly FileInfo path;            // path of the original file, if it was individual file (used to construct source file path)
-
-            internal JarItem(string name, byte[] data, FileInfo path)
-            {
-                this.name = name;
-                this.data = data;
-                this.path = path;
-            }
-        }
-
-        public struct Item
-        {
-            internal readonly Jar Jar;
-            private readonly int Index;
-
-            internal Item(Jar jar, int index)
-            {
-                this.Jar = jar;
-                this.Index = index;
-            }
-
-            internal string Name
-            {
-                get { return Jar.Items[Index].name; }
-            }
-
-            internal byte[] GetData()
-            {
-                return Jar.Items[Index].data;
-            }
-
-            internal FileInfo Path
-            {
-                get { return Jar.Items[Index].path; }
-            }
-
-            internal void Remove()
-            {
-                Jar.Items[Index] = new JarItem();
-            }
-
-            internal void MarkAsStub()
-            {
-                Jar.Items[Index] = new JarItem(Jar.Items[Index].name, null, null);
-            }
-
-            internal bool IsStub
-            {
-                get { return Jar.Items[Index].data == null; }
-            }
-        }
-
-        internal struct JarEnumerator
-        {
-            private readonly Jar jar;
-            private int index;
-
-            internal JarEnumerator(Jar jar)
-            {
-                this.jar = jar;
-                this.index = -1;
-            }
-
-            public Item Current
-            {
-                get { return new Item(jar, index); }
-            }
-
-            public bool MoveNext()
-            {
-                if (index + 1 < jar.Items.Count)
-                {
-                    index++;
-                    return true;
-                }
-                return false;
-            }
-        }
-
-        public JarEnumerator GetEnumerator()
-        {
-            return new JarEnumerator(this);
-        }
-    }
-
-    sealed class CompilerOptions
-    {
-        internal List<Jar> jars = new List<Jar>();
-        private Dictionary<string, int> jarMap = new Dictionary<string, int>();
-        internal int classesJar = -1;
-        internal int resourcesJar = -1;
-        internal bool nojarstubs;
-        internal FileInfo path;
-        internal FileInfo keyfile;
-        internal string keycontainer;
-        internal bool delaysign;
-        internal byte[] publicKey;
-        internal StrongNameKeyPair keyPair;
-        internal Version version;
-        internal string fileversion;
-        internal FileInfo iconfile;
-        internal FileInfo manifestFile;
-        internal bool targetIsModule;
-        internal string assembly;
-        internal string mainClass;
-        internal ApartmentState apartment;
-        internal PEFileKinds target;
-        internal bool guessFileKind;
-        internal string[] unresolvedReferences; // only used during command line parsing
-        internal Dictionary<string, string> legacyStubReferences = new Dictionary<string, string>();    // only used during command line parsing
-        internal Assembly[] references;
-        internal string[] peerReferences;
-        internal bool crossReferenceAllPeers = true;
-        internal string[] classesToExclude;     // only used during command line parsing
-        internal FileInfo remapfile;
-        internal Dictionary<string, string> props;
-        internal bool noglobbing;
-        internal CodeGenOptions codegenoptions;
-        internal DebugMode debugMode = DebugMode.Portable;
-        internal string debugFileName;
-        internal bool compressedResources;
-        internal string[] privatePackages;
-        internal string[] publicPackages;
-        internal string sourcepath;
-        internal Dictionary<string, string> externalResources;
-        internal string classLoader;
-        internal PortableExecutableKinds pekind = PortableExecutableKinds.ILOnly;
-        internal ImageFileMachine imageFileMachine = ImageFileMachine.I386;
-        internal ulong baseAddress;
-        internal uint fileAlignment;
-        internal bool highentropyva;
-        internal List<CompilerClassLoader> sharedclassloader; // should *not* be deep copied in Copy(), because we want the list of all compilers that share a class loader
-        internal HashSet<string> suppressWarnings = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        internal HashSet<string> errorWarnings = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        internal bool warnaserror; // treat all warnings as errors
-        internal FileInfo writeSuppressWarningsFile;
-        internal List<string> proxies = new List<string>();
-        internal object[] assemblyAttributeAnnotations;
-        internal bool warningLevelHigh;
-        internal bool noParameterReflection;
-        internal bool bootstrap;
-        internal string log;
-
-        internal CompilerOptions Copy()
-        {
-            CompilerOptions copy = (CompilerOptions)MemberwiseClone();
-            copy.jars = Copy(jars);
-            copy.jarMap = new Dictionary<string, int>(jarMap);
-            if (props != null)
-            {
-                copy.props = new Dictionary<string, string>(props);
-            }
-            if (externalResources != null)
-            {
-                copy.externalResources = new Dictionary<string, string>(externalResources);
-            }
-            copy.suppressWarnings = new(suppressWarnings, StringComparer.OrdinalIgnoreCase);
-            copy.errorWarnings = new(errorWarnings, StringComparer.OrdinalIgnoreCase);
-            return copy;
-        }
-
-        private static List<Jar> Copy(List<Jar> jars)
-        {
-            List<Jar> newJars = new List<Jar>();
-            foreach (Jar jar in jars)
-            {
-                newJars.Add(jar.Copy());
-            }
-            return newJars;
-        }
-
-        internal Jar GetJar(string file)
-        {
-            int existingJar;
-            if (jarMap.TryGetValue(file, out existingJar))
-            {
-                return jars[existingJar];
-            }
-            jarMap.Add(file, jars.Count);
-            return CreateJar(Path.GetFileName(file));
-        }
-
-        private Jar CreateJar(string jarName)
-        {
-            int count = 0;
-            string name = jarName;
-        retry:
-            foreach (Jar jar in jars)
-            {
-                if (jar.Name == name)
-                {
-                    name = Path.GetFileNameWithoutExtension(jarName) + "-" + (++count) + Path.GetExtension(jarName);
-                    goto retry;
-                }
-            }
-            Jar newJar = new Jar(name);
-            jars.Add(newJar);
-            return newJar;
-        }
-
-        internal Jar GetClassesJar()
-        {
-            if (classesJar == -1)
-            {
-                classesJar = jars.Count;
-                CreateJar("classes.jar");
-            }
-            return jars[classesJar];
-        }
-
-        internal bool IsClassesJar(Jar jar)
-        {
-            return classesJar != -1 && jars[classesJar] == jar;
-        }
-
-        internal Jar GetResourcesJar()
-        {
-            if (resourcesJar == -1)
-            {
-                resourcesJar = jars.Count;
-                CreateJar("resources.jar");
-            }
-
-            return jars[resourcesJar];
-        }
-
-        internal bool IsResourcesJar(Jar jar)
-        {
-            return resourcesJar != -1 && jars[resourcesJar] == jar;
-        }
-
-        internal bool IsExcludedClass(string className)
-        {
-            if (classesToExclude != null)
-            {
-                for (int i = 0; i < classesToExclude.Length; i++)
-                {
-                    if (Regex.IsMatch(className, classesToExclude[i]))
-                    {
-                        return true;
-                    }
-                }
-            }
-            return false;
-        }
-    }
-
-    class ManagedResolver : IManagedTypeResolver
-    {
-
-        readonly StaticCompiler compiler;
-
-        /// <summary>
-        /// Initializes a new instance.
-        /// </summary>
-        /// <param name="compiler"></param>
-        public ManagedResolver(StaticCompiler compiler)
-        {
-            this.compiler = compiler ?? throw new ArgumentNullException(nameof(compiler));
-        }
-
-        public Assembly ResolveBaseAssembly()
-        {
-            return compiler.baseAssembly;
-        }
-
-        public Assembly ResolveAssembly(string assemblyName)
-        {
-            return compiler.Universe.Load(assemblyName);
-        }
-
-        public Type ResolveCoreType(string typeName)
-        {
-            foreach (var assembly in compiler.Universe.GetAssemblies())
-                if (assembly.GetType(typeName) is Type t)
-                    return t;
-
-            return null;
-        }
-
-        public Type ResolveRuntimeType(string typeName)
-        {
-            return compiler.GetRuntimeType(typeName);
-        }
-
-    }
-
 }
