+++ src/IKVM.Tools.Importer/ImportArgLevel.cs
@@ -1,0 +1,26 @@
+﻿using System.Collections.Generic;
+
+namespace IKVM.Tools.Importer
+{
+
+    struct ImportArgLevel
+    {
+
+        public int Depth;
+        public List<string> Args;
+        public List<ImportArgLevel> Nested;
+
+        /// <summary>
+        /// Initializes a new instance.
+        /// </summary>
+        /// <param name="depth"></param>
+        public ImportArgLevel(int depth)
+        {
+            Depth = depth;
+            Args = new List<string>();
+            Nested = new List<ImportArgLevel>();
+        }
+
+    }
+
+}
