--- src/IKVM.Tools.Importer/ImportExternalResourceItem.cs
@@ -1,6 +1,0 @@
-﻿namespace IKVM.Tools.Importer
-{
-    public class ImportExternalResourceItem
-    {
-    }
-}
