--- src/IKVM.Tools.Importer/ImportResourceItem.cs
@@ -1,6 +1,0 @@
-﻿namespace IKVM.Tools.Importer
-{
-    internal class ImportResourceItem
-    {
-    }
-}
